[TECH-547] Code fix in consumer event handling

+++ book.query/src/main/java/com/happystays/book/query/domain/Supplier.java
@@ -15,7 +15,7 @@ import javax.persistence.*;
 public class Supplier {
     @Id
     @Column(name = "supplier_id")
-    private int supplierId;
+    private long supplierId;
     private String supplierCode;
     private String supplierName;
     private String supplierUri;
